Single queries file
