--- src/newHome/components/HomeActivities/queries.ts
@@ -1,28 +1,0 @@
-import { gql } from "@apollo/client";
-
-// TODO: Remove this query to homeActivities
-export const newHomeActivities = gql`
-  query HomeActivities($hasPermissionToManageOrders: Boolean!) {
-    activities: homepageEvents(last: 10) @include(if: $hasPermissionToManageOrders) {
-      edges {
-        node {
-          amount
-          composedId
-          date
-          email
-          emailType
-          id
-          message
-          orderNumber
-          oversoldItems
-          quantity
-          type
-          user {
-            id
-            email
-          }
-        }
-      }
-    }
-  }
-`;
