--- src/newHome/components/HomeSalesAnalytics/queries.tsx
@@ -1,13 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const queries = gql`
-  query HomeAnalitics($channel: String!, $hasPermissionToManageOrders: Boolean!) {
-    salesToday: ordersTotal(period: TODAY, channel: $channel)
-      @include(if: $hasPermissionToManageOrders) {
-      gross {
-        amount
-        currency
-      }
-    }
-  }
-`;
