+++ src/newHome/queries.ts
@@ -1,0 +1,50 @@
+import { gql } from "@apollo/client";
+
+// TODO: Remove this query to homeActivities
+export const newHomeActivities = gql`
+  query HomeActivities($hasPermissionToManageOrders: Boolean!) {
+    activities: homepageEvents(last: 10) @include(if: $hasPermissionToManageOrders) {
+      edges {
+        node {
+          amount
+          composedId
+          date
+          email
+          emailType
+          id
+          message
+          orderNumber
+          oversoldItems
+          quantity
+          type
+          user {
+            id
+            email
+          }
+        }
+      }
+    }
+  }
+`;
+
+// TODO: Remove this query to homeAnalytics
+export const newHomeAnalytics = gql`
+  query HomeAnaltics($channel: String!, $hasPermissionToManageOrders: Boolean!) {
+    salesToday: ordersTotal(period: TODAY, channel: $channel)
+      @include(if: $hasPermissionToManageOrders) {
+      gross {
+        amount
+        currency
+      }
+    }
+  }
+`;
+
+// TODO: Remove this query to homeNotifications
+export const newHomeNotifications = gql`
+  query homeNotifications($channel: String!) {
+    productsOutOfStock: products(filter: { stockAvailability: OUT_OF_STOCK }, channel: $channel) {
+      totalCount
+    }
+  }
+`;
